+++ ecmascript/compiler/codegen/maple/maple_be/src/cg/aarch64/aarch64_ebo.cpp
@@ -467,7 +467,7 @@ int32 AArch64Ebo::GetOffsetVal(const MemOperand &memOpnd) const
         val += static_cast<int>(offset->GetOffsetValue());
 
         if (offset->IsSymOffset() || offset->IsSymAndImmOffset()) {
-            val += offset->GetSymbol()->GetStIdx().Idx();
+            val += static_cast<int>(offset->GetSymbol()->GetStIdx().Idx());
         }
     }
     return val;
